Added a licenseToSign check
This constraint checks whether a license is signed before it is being saved. This prevents accidental saving of the same license without signing it first since a new license requires to be signed before saving. In case of a license loaded from a file, signing is required only when you add or change a feature